server/pyproject.toml
[project]
name = "raily-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.134.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.41.0",
]
